Make todo removable by the x button
